final 1

+++ KarnelTravel/Models/FlightInvoice.cs
@@ -9,19 +9,19 @@ public partial class FlightInvoice
 
     public int BookingId { get; set; }
 
-    public float FlightPrice { get; set; }
+    public double FlightPrice { get; set; }
 
     public int NumOfPassengers { get; set; }
 
-    public float SubTotal { get; set; }
+    public double SubTotal { get; set; }
 
-    public float Tax { get; set; }
+    public double Tax { get; set; }
 
     public string? DiscountCode { get; set; }
 
-    public float? DiscountPercent { get; set; }
+    public double? DiscountPercent { get; set; }
 
-    public float Total { get; set; }
+    public double Total { get; set; }
 
     public virtual Booking Booking { get; set; } = null!;
 }
